Complete fix for #33
